vcl: add "color channel bitcount" for an bitmap buffer (Linux)
In addition to "bitcount" on bitmap buffer, which represents the
number of bits one pixel takes, we also need the bitcount that the
color channels take of one pixel take.

For example RGB can be contined in 32 bits (because that the
underlaying buffer supports) In this case GetBitCount returns
32bit so we can traverse the buffer correctly. However, we also
need to know the bitcount of the actual color channels - in this
case this would be 24 bits.

Change-Id: I0a64928336540dc66d60659c46d64f369e786c3a