Support native alpha in outdev / virdev + support in headless
This adds support for native alpha surface (RGBA) to OutputDevice
and VirtualDevice. Mainly this removes restrictions that set the
alpha value to opaque (1.0f). All the functions still use a
separate "alpha" device for now.

As the headless backend already uses a 32-bit surface there was
not a big change necessary to make it work.

Change-Id: Ic2f2319a006764395d70b64c473749cd6d47f812